requirements.txt
pony
pandas
strawberry-graphql>=0.57.2
pyyaml
sqlalchemy
mysql-connector-python
pymysql
